Spec works + test

+++ CantinaApp.Core/ApplicationServices/Services/SpecialOffersServices.cs
@@ -18,16 +18,35 @@ namespace CantinaApp.Core.ApplicationServices.Services
 
         public SpecialOffers AddSpecialOffer(SpecialOffers specialOffers)
         {
+            
+            if (specialOffers.Price < 1)
+            {
+                throw new ArgumentException("Price need to be higher than 0");
+            }
+            if (specialOffers.SpecialOfferName == null)
+            {
+                throw new ArgumentException("Speciel offer need a name");
+            }
             return _sOffersRepo.CreateSpecialOffers(specialOffers);
         }
 
         public SpecialOffers GetSpecialOffersById(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException("Id need to be higher than 0");
+            }
+
             return _sOffersRepo.ReadSpecialOffers().ToList().FirstOrDefault(spc => spc.Id == id);
         }
 
         public SpecialOffers DeleteSpecialOffer(int id)
         {
+
+            if (id < 1)
+            {
+                throw new ArgumentException("Id need to be higher than 0");
+            }
             return _sOffersRepo.DeleteSpecialOffers(id);
         }
 
@@ -38,6 +57,18 @@ namespace CantinaApp.Core.ApplicationServices.Services
 
         public SpecialOffers UpdateSpecialOffer(SpecialOffers specialOffers)
         {
+            if (specialOffers.Price < 1)
+            {
+                throw new ArgumentException("Price need to be higher than 0");
+            }
+            if (specialOffers.SpecialOfferName == null)
+            {
+                throw new ArgumentException("Speciel offer need a name");
+            }
+            if (specialOffers.Id < 1)
+            {
+                throw new ArgumentException("Id need to be higher than 0");
+            }
             return _sOffersRepo.UpdateSpecialOffers(specialOffers);
         }
     }
